requirements.txt
numpy
matplotlib
scipy
torch
torchvision
pillow
tensorflow
sklearn
keras
h5py
